Tienda FUNCIONAL
He creado los 11 primeros generadores con sus respectivos puntos por segundo y precios, les he metido a todos los parámetros nuevos necesarios para su funcionamiento, he ajustado la tienda al cuadro layout, he creado el sistema de puntos por segundo y lo he hecho funcional y bonito. Ahora en la tienda cuando no has comprado el objeto anterior este aparece oculto.

+++ game/FeedTheLeader/Assets/Scripts/BonusPoints.cs
@@ -9,7 +9,7 @@ public class BonusPoints : MonoBehaviour
     public void OnMouseDown()
     {
         // Esta función se ejecuta cuando se hace clic en el objetivo
-        int pointsToAdd = Mathf.RoundToInt(PointsManager.Instance.getPuntos() * 0.1f); // Calculamos el 10% de los puntos totales y lo pasamos a Int
+        int pointsToAdd = Mathf.RoundToInt((float)(PointsManager.Instance.getPuntos() * 0.1f)); // Calculamos el 10% de los puntos totales y lo pasamos a Int
         if (pointsToAdd < 10)
         {
             pointsToAdd = 10; // Mínimo sumar 10 puntos
